# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KodeTalent API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a new student profile (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/by-email/{email}:
    get:
      operationId: getStudentByEmail
      tags: [students]
      summary: Get student by email
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /students/{id}/dashboard:
    get:
      operationId: getStudentDashboard
      tags: [students]
      summary: Get full dashboard data for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"
  /students/{id}/wrapped:
    get:
      operationId: getStudentWrapped
      tags: [students]
      summary: Get career wrapped data for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Career wrapped data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerWrapped"
  /quests:
    get:
      operationId: listQuests
      tags: [quests]
      summary: List quests filtered by field and year
      parameters:
        - name: field
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of quests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quest"
  /students/{id}/quests:
    get:
      operationId: getStudentQuests
      tags: [quests]
      summary: Get quest progress for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of student quest statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentQuestStatus"
  /students/{id}/quests/{questId}/complete:
    post:
      operationId: completeQuest
      tags: [quests]
      summary: Mark a quest as complete and award XP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: questId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated student after XP award
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Student:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        college:
          type: string
        city:
          type: string
        year:
          type: integer
        field:
          type: string
        githubUrl:
          type: string
          nullable: true
        overallScore:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        streakCount:
          type: integer
        lastActiveDate:
          type: string
          nullable: true
        skills:
          type: object
          additionalProperties:
            type: integer
        isPro:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - college
        - city
        - year
        - field
        - overallScore
        - xp
        - level
        - streakCount
        - skills
        - isPro
        - createdAt
    CreateStudentBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        college:
          type: string
        city:
          type: string
        year:
          type: integer
        field:
          type: string
        githubUrl:
          type: string
          nullable: true
      required:
        - name
        - email
        - college
        - city
        - year
        - field
    UpdateStudentBody:
      type: object
      properties:
        name:
          type: string
        college:
          type: string
        city:
          type: string
        year:
          type: integer
        field:
          type: string
        githubUrl:
          type: string
          nullable: true
        isPro:
          type: boolean
    DashboardData:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/Student"
        todayQuest:
          $ref: "#/components/schemas/Quest"
          nullable: true
        questProgress:
          type: object
          properties:
            completed:
              type: integer
            total:
              type: integer
          required:
            - completed
            - total
        topSkills:
          type: array
          items:
            $ref: "#/components/schemas/SkillEntry"
        collegeRank:
          type: object
          properties:
            rank:
              type: integer
            total:
              type: integer
            percentile:
              type: integer
          required:
            - rank
            - total
            - percentile
        xpToNextLevel:
          type: integer
      required:
        - student
        - questProgress
        - topSkills
        - collegeRank
        - xpToNextLevel
    CareerWrapped:
      type: object
      properties:
        studentId:
          type: integer
        name:
          type: string
        college:
          type: string
        streakDays:
          type: integer
        xpGained:
          type: integer
        scoreStart:
          type: integer
        scoreEnd:
          type: integer
        indiaPercentile:
          type: integer
        month:
          type: string
      required:
        - studentId
        - name
        - college
        - streakDays
        - xpGained
        - scoreStart
        - scoreEnd
        - indiaPercentile
        - month
    SkillEntry:
      type: object
      properties:
        name:
          type: string
        score:
          type: integer
      required:
        - name
        - score
    Quest:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        field:
          type: string
        year:
          type: integer
        xpReward:
          type: integer
        minutes:
          type: integer
        whyItMatters:
          type: string
        howToDoIt:
          type: string
      required:
        - id
        - title
        - description
        - field
        - year
        - xpReward
        - minutes
    StudentQuestStatus:
      type: object
      properties:
        quest:
          $ref: "#/components/schemas/Quest"
        status:
          type: string
          enum: [not_started, in_progress, completed]
        completedAt:
          type: string
          nullable: true
      required:
        - quest
        - status
...